requirements.txt
django==4.2.4
django-bootstrap4==23.2
django-pdfkit==0.3.1
djangorestframework==3.14.0
mysqlclient==2.2.0
pillow==10.0.0
pip==23.2.1
python==3.11.4